+++ app/src/main/java/com/and04/naturealbum/di/AuthenticationModule.kt
@@ -1,6 +1,7 @@
 package com.and04.naturealbum.di
 
-import com.and04.naturealbum.data.repository.FireBaseRepository
+import com.and04.naturealbum.data.repository.firebase.FireBaseRepository
+import com.and04.naturealbum.data.repository.firebase.UserRepository
 import com.and04.naturealbum.ui.mypage.AuthenticationManager
 import com.and04.naturealbum.ui.mypage.UserManager
 import dagger.Module
@@ -15,8 +16,8 @@ object AuthenticationModule {
     @Singleton
     @Provides
     fun providerAuthenticationManager(
-        fireBaseRepository: FireBaseRepository
-    ): AuthenticationManager = AuthenticationManager(fireBaseRepository)
+        userRepository: UserRepository
+    ): AuthenticationManager = AuthenticationManager(userRepository)
 
     @Provides
     fun providerUserManager() = UserManager
