+++ app/src/main/java/com/and04/naturealbum/di/FirebaseModule.kt
@@ -1,9 +1,11 @@
 package com.and04.naturealbum.di
 
-import com.and04.naturealbum.data.repository.FireBaseRepository
-import com.and04.naturealbum.data.repository.FireBaseRepositoryImpl
+import com.and04.naturealbum.data.datasource.FirebaseDataSource
+import com.and04.naturealbum.data.repository.firebase.FireBaseRepository
+import com.and04.naturealbum.data.repository.firebase.FireBaseRepositoryImpl
+import com.and04.naturealbum.data.repository.firebase.FriendRepository
+import com.and04.naturealbum.data.repository.firebase.UserRepository
 import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.firestore
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.FirebaseStorage
@@ -27,8 +29,26 @@ object FirebaseModule {
 
     @Provides
     @Singleton
-    fun providerFireBaseRepository(
+    fun providerFirebaseDataSource(
         fireStore: FirebaseFirestore,
         fireStorage: FirebaseStorage
-    ): FireBaseRepository = FireBaseRepositoryImpl(fireStore, fireStorage)
+    ): FirebaseDataSource = FirebaseDataSource(fireStore, fireStorage)
+
+    @Provides
+    @Singleton
+    fun providerFireBaseRepository(
+        firebaseDataSource: FirebaseDataSource
+    ): FireBaseRepository = FireBaseRepositoryImpl(firebaseDataSource)
+
+    @Provides
+    @Singleton
+    fun providerFriendRepository(
+        firebaseDataSource: FirebaseDataSource
+    ): FriendRepository = FriendRepository(firebaseDataSource)
+
+    @Provides
+    @Singleton
+    fun providerUserRepository(
+        firebaseDataSource: FirebaseDataSource
+    ): UserRepository = UserRepository(firebaseDataSource)
 }
