+++ app/src/main/java/com/and04/naturealbum/di/RoomModule.kt
@@ -3,12 +3,10 @@ package com.and04.naturealbum.di
 import android.content.Context
 import com.and04.naturealbum.data.repository.DataRepository
 import com.and04.naturealbum.data.repository.DataRepositoryImpl
-import com.and04.naturealbum.data.repository.FireBaseRepository
 import com.and04.naturealbum.data.room.AlbumDao
 import com.and04.naturealbum.data.room.AppDatabase
 import com.and04.naturealbum.data.room.LabelDao
 import com.and04.naturealbum.data.room.PhotoDetailDao
-import com.and04.naturealbum.ui.mypage.AuthenticationManager
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
