+++ app/src/main/java/com/and04/naturealbum/ui/friend/FriendViewModel.kt
@@ -6,7 +6,7 @@ import com.and04.naturealbum.data.dto.FirebaseFriend
 import com.and04.naturealbum.data.dto.FirebaseFriendRequest
 import com.and04.naturealbum.data.dto.FirestoreUserWithStatus
 import com.and04.naturealbum.data.dto.FriendStatus
-import com.and04.naturealbum.data.repository.FireBaseRepository
+import com.and04.naturealbum.data.repository.firebase.FriendRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -21,7 +21,7 @@ import javax.inject.Inject
 @OptIn(FlowPreview::class)
 @HiltViewModel
 class FriendViewModel @Inject constructor(
-    private val fireBaseRepository: FireBaseRepository,
+    private val friendRepository: FriendRepository,
 ) : ViewModel() {
 
     private val _receivedFriendRequests = MutableStateFlow<List<FirebaseFriendRequest>>(emptyList())
@@ -66,7 +66,7 @@ class FriendViewModel @Inject constructor(
 
     private fun fetchFilteredUsersAsFlow(currentUid: String, query: String) {
         viewModelScope.launch {
-            fireBaseRepository.searchUsersAsFlow(currentUid, query).collectLatest { results ->
+            friendRepository.searchUsersAsFlow(currentUid, query).collectLatest { results ->
                 _searchResults.value = results
             }
         }
@@ -75,7 +75,7 @@ class FriendViewModel @Inject constructor(
     private fun listenToFriends() {
         uid?.let { currentUid ->
             viewModelScope.launch {
-                fireBaseRepository.getFriendsAsFlow(currentUid).collect { friends ->
+                friendRepository.getFriendsAsFlow(currentUid).collect { friends ->
                     _friends.value = friends
                 }
             }
@@ -85,7 +85,7 @@ class FriendViewModel @Inject constructor(
     private fun listenToReceivedFriendRequests() {
         uid?.let { currentUid ->
             viewModelScope.launch {
-                fireBaseRepository.getReceivedFriendRequestsAsFlow(currentUid)
+                friendRepository.getReceivedFriendRequestsAsFlow(currentUid)
                     .collect { receivedFriendRequests ->
                         _receivedFriendRequests.value = receivedFriendRequests
                     }
@@ -95,7 +95,7 @@ class FriendViewModel @Inject constructor(
 
     fun sendFriendRequest(uid: String, targetUid: String) {
         viewModelScope.launch {
-            val success = fireBaseRepository.sendFriendRequest(uid, targetUid)
+            val success = friendRepository.sendFriendRequest(uid, targetUid)
             if (success) {
                 // 친구 요청이 성공적으로 전송되었을 경우 UI 상태를 업데이트
                 _searchResults.value = _searchResults.value.toMutableMap().apply {
@@ -109,13 +109,13 @@ class FriendViewModel @Inject constructor(
 
     fun acceptFriendRequest(uid: String, targetUid: String) {
         viewModelScope.launch {
-            fireBaseRepository.acceptFriendRequest(uid, targetUid)
+            friendRepository.acceptFriendRequest(uid, targetUid)
         }
     }
 
     fun rejectFriendRequest(uid: String, targetUid: String) {
         viewModelScope.launch {
-            fireBaseRepository.rejectFriendRequest(uid, targetUid)
+            friendRepository.rejectFriendRequest(uid, targetUid)
         }
     }
 }
